# ARROW-3643: [Rust] optimize BooleanBufferBuilder::append_slice