Performance difference: ~15%

```
bench_bool              time:   [2.6103 ms 2.6278 ms 2.6474 ms]                        
                        thrpt:  [188.86 MiB/s 190.27 MiB/s 191.55 MiB/s]
                 change:
                        time:   [-16.950% -15.811% -14.657%] (p = 0.00 < 0.05)
                        thrpt:  [+17.174% +18.781% +20.410%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

Travis-CI: https://travis-ci.com/lihalite/arrow/builds/126254420